[Flarm5] Driver/FLARM - catch all std exceptions in the reading of flight information
in function ReadFlightList (for example if the duration of response to large).

In this case, not only is the flight in question not displayed in the list,
but the entire loop is exited, which means that the flight list is completely empty at the end
although some flights have already been detected!

August2111: mache 3 versuche beim Info herauslesen,
Progressbar sollte den Wert der max. Flüge haben
3 Versuche beim FlightInfo read - das in die richtige Reihenfolge bringen!